# infra/gateway/openapi_v2.yaml
swagger: "2.0"
info: { title: "LIT Gateway — Search Unified", version: "v2" }
schemes: ["https"]
produces: ["application/json"]

paths:
  /public/status:
    options:
      operationId: cors_status
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      responses: { "200": { description: "ok" } }
    get:
      operationId: status
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      responses: { "200": { description: "ok" } }

  /public/getFilterOptions:
    options:
      operationId: cors_getFilterOptions
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      responses: { "200": { description: "ok" } }
    get:
      operationId: getFilterOptions
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      responses: { "200": { description: "ok" } }

  /public/searchCompanies:
    options:
      operationId: cors_searchCompanies
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      responses: { "200": { description: "ok" } }
    post:
      operationId: searchCompanies
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              q:        { type: string }
              page:     { type: integer }
              pageSize: { type: integer }
      responses: { "200": { description: "ok" } }

  /public/getCompanyShipments:
    options:
      operationId: cors_getCompanyShipments
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      responses: { "200": { description: "ok" } }
    get:
      operationId: getCompanyShipments
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      parameters:
        - { name: company_id, in: query, type: string, required: true }
        - { name: limit,      in: query, type: integer, required: false }
        - { name: offset,     in: query, type: integer, required: false }
      responses: { "200": { description: "ok" } }

  /public/iy/searchShippers:
    options:
      operationId: cors_iySearchShippers
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      responses: { "200": { description: "ok" } }
    post:
      operationId: iySearchShippers
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              q:        { type: string }
              page:     { type: integer }
              pageSize: { type: integer }
      responses: { "200": { description: "ok" } }

  /public/iy/companyBols:
    options:
      operationId: cors_iyCompanyBols
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      responses: { "200": { description: "ok" } }
    get:
      operationId: iyCompanyBols
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      parameters:
        - { name: company_id, in: query, type: string,  required: true }
        - { name: limit,      in: query, type: integer, required: false }
        - { name: offset,     in: query, type: integer, required: false }
      responses: { "200": { description: "ok" } }

  /public/iy/companyProfile:
    get:
      summary: Get ImportYeti company profile (normalized)
      operationId: getIyCompanyProfile
      parameters:
        - name: company_id
          in: query
          required: true
          type: string
          description: "Company key, e.g. company/walmart"
      responses:
        "200":
          description: Company profile
          schema:
            type: object
        "400":
          description: Upstream ImportYeti error (auth/credits/etc.)
        "500":
          description: Internal error
      x-google-backend:
        address: __SERVICE_URL__
        jwt_audience: __SERVICE_URL__
      x-google-cors:
        allowOrigins: ["*"]
        allowMethods: ["GET", "OPTIONS"]
        allowHeaders: ["content-type", "authorization", "x-api-key"]

  /public/iy/companyLatestBol:
    options:
      operationId: cors_iyCompanyLatestBol
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      responses: { "200": { description: "ok" } }
    get:
      operationId: iyCompanyLatestBol
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      parameters:
        - { name: company, in: query, type: string, required: true }
      responses: { "200": { description: "ok" } }

  /public/iy/companyStats:
    options:
      operationId: cors_iyCompanyStats
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      responses: { "200": { description: "ok" } }
    get:
      operationId: iyCompanyStats
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      parameters:
        - { name: company, in: query, type: string, required: true }
        - { name: range, in: query, type: string, required: false }
      responses: { "200": { description: "ok" } }

  /crm/saveCompany:
    options:
      operationId: cors_crmSaveCompany
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      responses: { "200": { description: "ok" } }
    post:
      operationId: crmSaveCompany
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              company_id: { type: string }
              stage:      { type: string }
              provider:   { type: string }
              payload:    { type: object }
      responses: { "200": { description: "ok" } }

  /crm/savedCompanies:
    options:
      operationId: cors_crmSavedCompanies
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      responses: { "200": { description: "ok" } }
    get:
      operationId: crmSavedCompanies
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      parameters:
        - name: stage
          in: query
          required: false
          type: string
      responses: { "200": { description: "ok" } }

  /crm/companies/{company_id}:
    options:
      operationId: cors_crmGetCompany
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      responses: { "200": { description: "ok" } }
    get:
      operationId: crmGetCompany
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      parameters:
        - name: company_id
          in: path
          required: true
          type: string
      responses: { "200": { description: "ok" } }

  /crm/companies/{company_id}/enrichContacts:
    options:
      operationId: cors_crmEnrichContacts
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      responses: { "200": { description: "ok" } }
    post:
      operationId: crmEnrichContacts
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      parameters:
        - name: company_id
          in: path
          required: true
          type: string
      responses: { "200": { description: "ok" } }

  /crm/campaigns:
    options:
      operationId: cors_crmCampaigns
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      responses: { "200": { description: "ok" } }
    get:
      operationId: crmGetCampaigns
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      responses: { "200": { description: "ok" } }
    post:
      operationId: crmCreateCampaign
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              name: { type: string }
              sequence: { type: object }
              settings: { type: object }
      responses: { "200": { description: "ok" } }

  /crm/campaigns/{campaign_id}:
    options:
      operationId: cors_crmGetCampaign
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      responses: { "200": { description: "ok" } }
    get:
      operationId: crmGetCampaign
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      parameters:
        - name: campaign_id
          in: path
          required: true
          type: integer
      responses: { "200": { description: "ok" } }

  /crm/campaigns/{campaign_id}/addCompany:
    options:
      operationId: cors_crmAddCompanyToCampaign
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      responses: { "200": { description: "ok" } }
    post:
      operationId: crmAddCompanyToCampaign
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      parameters:
        - name: campaign_id
          in: path
          required: true
          type: integer
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              company_id: { type: string }
              contact_ids: { type: array, items: { type: string } }
      responses: { "200": { description: "ok" } }

  /rfp/company/{company_id}/context:
    options:
      operationId: cors_rfpGetContext
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      responses: { "200": { description: "ok" } }
    get:
      operationId: rfpGetContext
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      parameters:
        - name: company_id
          in: path
          required: true
          type: string
      responses: { "200": { description: "ok" } }

  /rfp/generate:
    options:
      operationId: cors_rfpGenerate
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      responses: { "200": { description: "ok" } }
    post:
      operationId: rfpGenerate
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              company_id: { type: string }
              lanes: { type: array, items: { type: object } }
              owner: { type: string }
              template: { type: string }
      responses: { "200": { description: "ok" } }

  /rfp/workspace:
    options:
      operationId: cors_rfpWorkspace
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      responses: { "200": { description: "ok" } }
    post:
      operationId: rfpWorkspace
      security: []
      x-google-backend: { address: __SERVICE_URL__, jwt_audience: __SERVICE_URL__ }
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              company_id: { type: string }
              name: { type: string }
              lanes: { type: array, items: { type: object } }
      responses: { "200": { description: "ok" } }